rccn/sms: Remove unused send immediate command and vty dependency
The string "send_immediate" does not appear in the rccn or puppet
repositories except for here per recursive grep. Removing the
send_immediate method (which appears to be unused) allows for the SMS
module to no longer depend directly on the obscvty module. This limits
the surface area of the code exposed to the VTY interface which needs
to be migrated to support the split stack instead of the NITB.