rccn: Mark get_imsi as a public method
Python is fully dynamic and has no real private methods, but by
convention functions which begin with underscores are considered
private and should not be accessed externally (with some rare
exceptions like in named_tuple). The `_get_imsi(msisdn)` method was used
externally to the `Subscriber` class in rrc.py. To better follow
convention this commit changes this to a public method since it is
used publicly.