rccn/subscriber: Update edit implementation to avoid introspection
The edit implementation previously coupled the database column names
to the function parameter names. This commit explicitly sets the
mapping, making it more explicit which columns are modified. This
allows for the column names to change in the future without requiring
a change to the high-level internal api as well.