rccn/rrc: Maintain existing behavior with extension api
The rrc used a special subscriber api which returned an extension from
an IMSI if the subscriber was attached, and would otherwise throw an
exception. To limit the subscriber api surface, this commit updates
RRC to achieve this same behavior with existing subscriber apis. This
makes the logic more explicit where it is relevant while helping to
limit the subscriber API surface area.