rccn/subscriber: Consistently close local db connections
The connection close logic was implemented inconsistently, and would
leak some connection objects. This is not a big deal since the
connections are relatively lightweight, but can result in extra
resource usage when the system is loaded. This was picked up by the
pycharm db linter. This commit uses the try/except/finally syntax
provided by python to make clear that the connection will always be
closed.