rccn & bin: Update SMS db naming to separate from hlr
In the split osmocom components the SMS table is generated from the
smsc subcomponent of the msc, and stored in a separate sqlite3 table
from the hlr. This commit updates accesses to the SMS table in the
smsc to make it clear it is not necessarily the same as the overall
hlr database.

This commit requires an update to the config.py template in puppet to
include the sms_db parameter before deployment.