rccn/osmonitb + hlr: Lookup IMSI with VTY instead of sqlite
The IMSI lookups happen frequently in some components, and excessive
sqlite use has been linked to corruption of the sqlite database at
times in the past. The VTY is a safer interface, even if somewhat less
elegant.

Use the VTY for now until a GSUP or ctrl-based interface can be
created.